Add 'TurnOffAndOnLight' to '/twitch/rewards'
