+++ src/index.ts
@@ -1,5 +1,10 @@
 import { Request, ResponseToolkit, Server } from "@hapi/hapi";
+import { NodeFetchHttpDriver } from "./drivers/NodeFetchHttpDriver";
+import { Light } from "./entities/Light";
+import { Reward } from "./entities/Reward";
+import { LifxLightService } from "./services/light/LifxLightService";
 import { initTwitchPubSub } from "./twitch";
+import { TurnOffAndOnLights } from "./useCases/TurnOffAndOnLights";
 
 const init = async () => {
   const server = new Server({
@@ -18,8 +23,17 @@ const init = async () => {
       path: "/twitch/rewards",
       method: "POST",
       handler: async (request: Request, h: ResponseToolkit) => {
-        const payload = request.payload;
-        console.log(payload);
+        const { reward } = request.payload as { reward: Reward };
+        console.log(reward);
+
+        switch (reward.id) {
+          case "3cc420c0-2198-4cd8-9a07-c68f489c50be":
+            const lights: Light[] = (process.env.LIFX_LIGHTS || "")
+              .split(",")
+              .map((id) => ({ id, service: "LIFX" }));
+            await new TurnOffAndOnLights().perform({ lights });
+            break;
+        }
 
         return h.response().code(204);
       },
@@ -27,7 +41,7 @@ const init = async () => {
   ]);
 
   await server.start();
-  initTwitchPubSub();
+  // initTwitchPubSub();
   console.log("Server running on %s", server.info.uri);
 };
